Treat \9 legacy IE property hack as a vendor extension warning (mostly)
Fixes part of #10.
When "Treat Vendor Extensions as warnings" is enabled, most usages of
the \9 hack will result in warnings instead of errors.

This addresses the following:
    prop: value \9;
    prop: value\9;

This commit unfortunately does not fix the specific case of:
    prop: «number»«unit?»\9;
Because that will require fiddling with the parser grammar (viz. the <DIMEN> token)