Add target to build jar with no dependency jars
The ant “jar” target in build.xml here now builds a css-validator.jar file that
includes all the dependency jars from the lib/ subdirectory.  But in some cases
it’s also still useful to have a css-validator.jar without the dependencies; so
this change adds an additional “jar-without-dependencies” target for doing that.