Add org.apache.velocity.io.UnicodeInputStream copy
This change adds a copy of the org.apache.velocity.io.UnicodeInputStream source
from the https://svn.apache.org/repos/asf/velocity/engine repo, changing the
package name to org.w3c.css.util, and updating ApplContext and HTTPURL to rely
on it instead of importing org.apache.velocity.io.UnicodeInputStream.

The rationale for this change is, it allows the CSS validator to be integrated
into other applications without any dependency on Apache Velocity (the jar
distro for which is 439K…). Specifically in the case of the Nu Html Checker, it
allows the CSS validator to be integrated without needing any new dependencies
at to be added to the existing Nu Html Checker dependency set.