Make css-validator.jar CLI runnable anywhere
This change makes it possible to run the css-validator.jar CLI from anywhere.

Prior to this change, the css-validator.jar CLI could be run only from within a
working directory that had the org/w3c/css/css/*.properties Velocity templates.
That’s because the code for the CLI looked for the templates on the filesystem,
or looked for a jar on the filesystem containing the templates.

This change makes the CLI code instead just look for the templates anywhere in
the classpath. And the templates are in the css-validator.jar file, so it works
regardless of whether you invoke the CLI with `java -jar css-validator.jar` or
instead with `java -cp css-validator.jar org.w3c.css.css.CssValidator`.

This change also makes the CLI stop looking for the dependency jar files on the
filesystem in a `lib/` subdirectory. In the case where you’re using the full
css-validator.jar file that contains all the dependency jars, there’s no need to
look for them on the filesystem. And in the case where you’re using the
css-validator.jar that doesn’t have the dependency jars included, you can just
call org.w3c.css.css.CssValidator with the jars in your classpath; for example:

  java -cp ~/css-validator/lib/* org.w3c.css.css.CssValidator https://www.w3.org

Note that this change replaces the code added in 6b7b237 and c6cbb04 that took
the approach of looking for the templates and lib/ dir on the filesystem.