Allow 'pointer-events: auto' but report warning
This change allows the `auto` value for the `pointer-events` property but
reports a warning to inform users that it’s not defined in any specifications.

The `auto` value for `pointer-events` is used commonly in the wild — even though
it’s not defined as valid in any CSS specs — because support for it has been
implemented in browsers for more than 10 years now.

Details are in comments in https://bugzilla.mozilla.org/show_bug.cgi?id=380573
and https://webkit.org/specs/PointerEventsProperty.html is the spec which the
browsers implemented from.