Prevent empty URI/upload/textarea input in Web UI
This change adds client-side form validation to the Web UI, to provide a better
user experience: Users get error messages for empty URI/upload/textarea input
directly in the Web UI before the frontend submits any request to the server.

The text used in the form-validation error message pop-ups/tooltips is copied
from the existing instructions in the (localized) HTML content fo the page. So
the form-validation error messages are shown in the appropriate user locale.