Treat the 'zoom' property as a vendor extension
This change causes the `zoom` property to be handled as a vendor extension if
the option to treat vendor extensions as warnings is set.

The change is implemented by providing a new hook in the code where the names of
any other commonly-used non-standard properties can be added later (if any other
than `zoom` ever end up being judged appropriate for similar special treatment).