Simplify internal CSS-version handling for colors
Internally the code was maintaining a list of “CSS4” colors and a list of “CSS3”
colors, but the only difference between the two was the color `rebeccapurple`.

So this change adds `rebeccapurple` to the list of “CSS3” colors and changes the
rest of the code to use that wherever it was previously using the “CSS4” list.

That old code was the only place where a “CSS4” distinction was used for any
feature being checked. So this change makes the color-checking version behavior
consistent with the rest of the code, where the “CSS3” label now just means in
practice “the latest relevant spec for whatever feature is being checked”.