Catch TokenMgrError in attributeDeclarations()
This change causes any TokenMgrError thrown in attributeDeclarations()
in the parser code to be caught and re-thrown as a ParseException so
that is gets reported as an error along with other errors.

Without this change, if parseStyleAttribute() is called from client code
on a case like one of the following:

    <div id="myGallery" style="background:#;width:100%;">
    <div .="" br="" div="" style="\">

...that is, with a syntax error, then a TokenMgrError is thrown but
never caught in any other part of the code. So it just becomes an
unhandled exception that without this change the client code would
otherwise have to write its own handling for.

The Web-based checker apparently has other code higher up for catching
that TokenMgrError and handling it. And of course handling for it could
be added to any other client code — but catching it earlier means fewer
unexpected surprises for anybody trying to write other client code.