Handle 1x, 2x, etc., as used in -webkit-image-set
This change makes the parser recognize dimensions of the form 1x, 2x,
etc. — as used by `-webkit-image-set` — while flagging those dimensions
a non-standard vendor extensions (so that specific vendor-extensions
warnings get reported for them if the user/client has opted to have such
cases reported as warnings instead of as errors).

Without this change, errors get reported for such dimension values
regardless of the context. The reason is the parser doesn’t recognize
those dimenstions at all, and because the parser tokenizes dimensions
before checking for vendor extensions, errors always get reported for
those instances even if they are part of the syntax of a non-standard
vendor extension such as `-webkit-image-set` that the backend code
otherwise already handles as a warning case.

That is, without this change there’s no way to tell the code to suppress
the error messages for those particular dimension instances because
they’re part of a vendor extension — because the parser code that checks
the dimension values executes before the code to checks whether the
expression that contains them is a vendor extension.

Addresses https://github.com/validator/validator/issues/607