Allow @supports and @viewport (without checking)
This change is a temporary workaround that adds an interim/initial/partial level
of support for allowing the @supports and @viewport at-rules. The intent is that
we’ll add full support for them later, and then at that time back this change
out (because adding real, full support will make it unnecessary).

This change simply causes no errors to be reported for stylesheets that use
@supports and @viewport. Beyond that, it doesn’t add any checking to report an
error if the content of a particular @supports at-rule or @viewport at-rule
doesn’t conform to the requirements in the relevant CSS specs.

Addresses https://github.com/w3c/css-validator/issues/71